исправлен int на long в croppedMiddle

+++ src/SalesManager.java
@@ -14,8 +14,8 @@ public class SalesManager {
         }
         return max;
     }
-    public int min() {
-        int min = sales[0];
+    public long min() {
+        long min = sales[0];
         for (int i = 0;i<sales.length;i++) {
             if (sales[i] >= 0 && sales[i] < min) {
                 min = sales[i];
@@ -23,13 +23,12 @@ public class SalesManager {
         }
         return min;
     }
-    public int croppedMiddle() {
-        int salesAmount = 0;
-        //int min = -1;
-        for (int sale : sales) {
+    public long croppedMiddle() {
+        long salesAmount = 0;
+        for (long sale : sales) {
             salesAmount += sale;
         }
-        int croppedMiddle = (salesAmount - max() - min())/(sales.length - 2);
+        long croppedMiddle = (salesAmount - max() - min())/(sales.length - 2);
         return croppedMiddle;
     }
 }
